+++ AnthroCloud/AnthroCloud.UI.Blazor/Services/ChartService.cs
@@ -16,7 +16,7 @@ namespace AnthroCloud.UI.Blazor.Services
             this.httpClient = httpClient;
         }
 
-        public async Task<string> GetAllBfaJson(byte id, byte x, decimal y, GraphTypes z)
+        public async Task<string> GetAllBfaJson(byte id, byte x, double y, GraphTypes z)
         {
             string uri = $"chart/BFAJson/{id}/{x}/{y}/{z}";
             var response = await httpClient.GetAsync(uri);
@@ -26,7 +26,7 @@ namespace AnthroCloud.UI.Blazor.Services
             return responseBody;
         }
 
-        public async Task<string> GetAllHcfaJson(byte id, byte x, decimal y, GraphTypes z)
+        public async Task<string> GetAllHcfaJson(byte id, byte x, double y, GraphTypes z)
         {
             string uri = $"chart/HCFAJson/{id}/{x}/{y}/{z}";
             var response = await httpClient.GetAsync(uri);
@@ -36,7 +36,7 @@ namespace AnthroCloud.UI.Blazor.Services
             return responseBody;
         }
 
-        public async Task<string> GetAllLhfaJson(byte id, byte x, decimal y, GraphTypes z)
+        public async Task<string> GetAllLhfaJson(byte id, byte x, double y, GraphTypes z)
         {
             string uri = $"chart/LHFAJson/{id}/{x}/{y}/{z}";
             var response = await httpClient.GetAsync(uri);
@@ -46,7 +46,7 @@ namespace AnthroCloud.UI.Blazor.Services
             return responseBody;
         }
 
-        public async Task<string> GetAllMuacJson(byte id, byte x, decimal y, GraphTypes z)
+        public async Task<string> GetAllMuacJson(byte id, byte x, double y, GraphTypes z)
         {
             string uri = $"chart/MUACJson/{id}/{x}/{y}/{z}";
             var response = await httpClient.GetAsync(uri);
@@ -56,7 +56,7 @@ namespace AnthroCloud.UI.Blazor.Services
             return responseBody;
         }
 
-        public async Task<string> GetAllSsfaJson(byte id, byte x, decimal y, GraphTypes z)
+        public async Task<string> GetAllSsfaJson(byte id, byte x, double y, GraphTypes z)
         {
             string uri = $"chart/SSFAJson/{id}/{x}/{y}/{z}";
             var response = await httpClient.GetAsync(uri);
@@ -66,7 +66,7 @@ namespace AnthroCloud.UI.Blazor.Services
             return responseBody;
         }
 
-        public async Task<string> GetAllTsfaJson(byte id, byte x, decimal y, GraphTypes z)
+        public async Task<string> GetAllTsfaJson(byte id, byte x, double y, GraphTypes z)
         {
             string uri = $"chart/TSFAJson/{id}/{x}/{y}/{z}";
             var response = await httpClient.GetAsync(uri);
@@ -76,7 +76,7 @@ namespace AnthroCloud.UI.Blazor.Services
             return responseBody;
         }
 
-        public async Task<string> GetAllWfaJson(byte id, byte x, decimal y, GraphTypes z)
+        public async Task<string> GetAllWfaJson(byte id, byte x, double y, GraphTypes z)
         {
             string uri = $"chart/WFAJson/{id}/{x}/{y}/{z}";
             var response = await httpClient.GetAsync(uri);
@@ -86,7 +86,7 @@ namespace AnthroCloud.UI.Blazor.Services
             return responseBody;
         }
 
-        public async Task<string> GetAllWfhJson(byte id, decimal x, decimal y, GraphTypes z)
+        public async Task<string> GetAllWfhJson(byte id, double x, double y, GraphTypes z)
         {
             string uri = $"chart/WFHJson/{id}/{x}/{y}/{z}";
             var response = await httpClient.GetAsync(uri);
@@ -96,7 +96,7 @@ namespace AnthroCloud.UI.Blazor.Services
             return responseBody;
         }
 
-        public async Task<string> GetAllWflJson(byte id, decimal x, decimal y, GraphTypes z)
+        public async Task<string> GetAllWflJson(byte id, double x, double y, GraphTypes z)
         {
             string uri = $"chart/WFLJson/{id}/{x}/{y}/{z}";
             var response = await httpClient.GetAsync(uri);
